+++ components/itemImage.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default class ItemImage extends React.Component {
-  render() {}
-}