+++ next.config.js
@@ -1,13 +1,27 @@
+const withBundleAnalyzer = require('@zeit/next-bundle-analyzer');
 const withCss = require('@zeit/next-css');
+const withOffline = require('next-offline');
 const withPlugins = require('next-compose-plugins');
+const withSourceMaps = require('@zeit/next-source-maps');
+
+const bundleAnalyzerConfig = {
+  analyzeServer: ['server', 'both'].includes(process.env.BUNDLE_ANALYZE),
+  analyzeBrowser: ['browser', 'both'].includes(process.env.BUNDLE_ANALYZE),
+  bundleAnalyzerConfig: {
+    server: {
+      analyzerMode: 'static',
+      reportFilename: '../../bundles/server.html',
+    },
+    browser: {
+      analyzerMode: 'static',
+      reportFilename: '../bundles/client.html',
+    },
+  },
+};
 
 const nextConfig = {
-  webpack: (config, { buildId, dev, isServer, defaultLoaders }) => {
-    return config;
-  },
-  webpackDevMiddleware: config => {
-    return config;
-  },
+  webpack: (config, { buildId, dev, isServer, defaultLoaders }) => config,
+  webpackDevMiddleware: config => config,
   publicRuntimeConfig: {
     hostname: process.env.NOW
       ? 'https://case-study.anthony.codes'
@@ -15,4 +29,12 @@ const nextConfig = {
   },
 };
 
-module.exports = withPlugins([withCss], nextConfig);
+module.exports = withPlugins(
+  [
+    withBundleAnalyzer(bundleAnalyzerConfig),
+    withCss,
+    withOffline,
+    withSourceMaps,
+  ],
+  nextConfig,
+);